requirements.txt
nikola
jinja2
aiohttp
watchdog
flask
loguru
pyyaml
waitress